fix warning

+++ modules/elementary_functions/builtin/cpp/ndimsBuiltin.cpp
@@ -62,7 +62,7 @@ ArrayOfVector Nelson::ElementaryFunctionsGateway::ndimsBuiltin(Evaluator* eval, 
             case NLS_CELL_ARRAY:
             case NLS_STRUCT_ARRAY:
             {
-                double ndims = param1.getDimensions().getLength();
+                double ndims = (double)param1.getDimensions().getLength();
                 if (ndims < 2)
                 {
                     ndims = 2;
